reverting some crap

+++ GithubPresentation/src/githubbbbb/CalculatorInterface.java
@@ -6,5 +6,5 @@ public interface CalculatorInterface {
 	int subtract(int num1, int num2);
 	int divide(int num1, int num2);
 	int multiply(int num1, int num2);
-	float modulo(float num1, float num2);
+	int modulo(int num1, int num2);
 }
